firmware: arm_scmi: Cleanup core driver removal callback
Platform drivers .remove callbacks are not supposed to fail and report
errors: such errors are indeed ignored by the core platform drivers stack
and the driver unbind process is anyway completed.

The SCMI core platform driver as it is now, instead, bails out reporting
an error in case of an explicit unbind request.

Fix the removal path by adding proper device links between the core SCMI
device and the SCMI protocol devices so as to cause a full SCMI stack
unbind when the core driver is removed: the remove process does not bail
out anymore on the anomalous conditions triggered by an explicit unbind
but the user is still warned.

Reported-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Cristian Marussi <cristian.marussi@arm.com>
Link: https://lore.kernel.org/r/20221028140833.280091-1-cristian.marussi@arm.com
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>